test: group SplitTooltip tests
